+++ terminol/xcb/window.cxx
@@ -803,8 +803,8 @@ bool Window::terminalFixDamageBegin(bool internal) throw () {
 
 void Window::terminalDrawRun(uint16_t        row,
                              uint16_t        col,
-                             uint8_t         fg,
-                             uint8_t         bg,
+                             uint16_t        fg,
+                             uint16_t        bg,
                              AttributeSet    attrs,
                              const uint8_t * str,
                              size_t          count) throw () {
@@ -844,8 +844,8 @@ void Window::terminalDrawRun(uint16_t        row,
 
 void Window::terminalDrawCursor(uint16_t        row,
                                 uint16_t        col,
-                                uint8_t         fg,
-                                uint8_t         bg,
+                                uint16_t        fg,
+                                uint16_t        bg,
                                 AttributeSet    attrs,
                                 const uint8_t * str,
                                 bool            special) throw () {
